created delete and patch endpoint

+++ fast_zero/routers/todos.py
@@ -1,10 +1,20 @@
+from http import HTTPStatus
+
 from fastapi import APIRouter, Depends
+from fastapi.exceptions import HTTPException
 from sqlalchemy import select
 from sqlalchemy.orm import Session
 
 from fast_zero.database import get_session
 from fast_zero.models import ToDo
-from fast_zero.schemas import ToDoList, ToDoPublic, ToDoSchema, ToDoState, Message
+from fast_zero.schemas import (
+    Message,
+    ToDoList,
+    ToDoPublic,
+    ToDoSchema,
+    ToDoState,
+    ToDoUpdate,
+)
 from fast_zero.security import get_current_user
 
 router = APIRouter(prefix='/todos', tags=['todos'])
@@ -55,18 +65,53 @@ def list_todos(  # noqa
 
     return {'todos': todos}
 
-@router.delete('/', response_model=Message)
+
+@router.delete('/{todo_id}', response_model=Message)
 def delete_todo(
-    todo_id: int = None,
+    todo_id: int,
     session: Session = Depends(get_session),
-    current_user = Depends(get_current_user)
+    current_user=Depends(get_current_user),
 ):
-    db_todo = session.scalar(select(ToDo).where(
-        (ToDo.id == todo_id) | (ToDo.user_id == current_user.id)
+    db_todo = session.scalar(
+        select(ToDo).where(
+            (ToDo.id == todo_id) & (ToDo.user_id == current_user.id)
         )
     )
+
+    if not db_todo:
+        raise HTTPException(
+            status_code=HTTPStatus.NOT_FOUND, detail='Task not found'
+        )
 
     session.delete(db_todo)
     session.commit()
 
     return {'message': 'Task deleted successfully'}
+
+
+@router.patch('/{todo_id}', response_model=ToDoPublic)
+def patch_todo(
+    todo_id: int,
+    todo: ToDoUpdate,
+    session: Session = Depends(get_session),
+    current_user=Depends(get_current_user),
+):
+    db_todo = session.scalar(
+        select(ToDo).where(
+            (ToDo.id == todo_id) & (ToDo.user_id == current_user.id)
+        )
+    )
+
+    if not db_todo:
+        raise HTTPException(
+            status_code=HTTPStatus.NOT_FOUND, detail='Task not found'
+        )
+
+    for key, value in todo.model_dump(exclude_unset=True).items():
+        setattr(db_todo, key, value)
+
+    session.add(db_todo)
+    session.commit()
+    session.refresh(db_todo)
+
+    return db_todo
